+++ src/main/java/agh/ics/proj1/Config.java
@@ -2,10 +2,10 @@ package agh.ics.proj1;
 
 public class Config {
     public final int mapWidth, mapHeight;
-    public final double grassEnergy;
-    public final double animalInitialEnergy;
+    public final int grassEnergy;
+    public final int animalInitialEnergy;
 
-    public Config(int mapWidth, int mapHeight, double grassEnergy, double animalInitialEnergy) {
+    public Config(int mapWidth, int mapHeight, int grassEnergy, int animalInitialEnergy) {
         this.mapWidth = mapWidth;
         this.mapHeight = mapHeight;
         this.grassEnergy = grassEnergy;
